pyproject.toml
[project]
name = "zeabur-django-serverless-demo"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "AngelLiang", email = "pl01665077@163.com" }
]
dependencies = [
    "django>=5.1.2",
    "django-ninja>=1.3.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []
